(#63, #338, #339, #340, #341) Code review changes
Clarify docs, change type visibility, fix namespaces, connect TODOs with
the issues.
