+++ src/WpfMath.Shared/Atoms/DummyAtom.cs
@@ -1,4 +1,5 @@
 using WpfMath.Boxes;
+using WpfMath.Fonts;
 using WpfMath.Utils;
 
 namespace WpfMath.Atoms
