+++ src/WpfMath.Shared/Atoms/SymbolAtom.cs
@@ -1,8 +1,11 @@
 using System;
 using System.Collections;
 using System.Collections.Generic;
-using System.Diagnostics.CodeAnalysis;
+using WpfMath.Fonts;
 using WpfMath.Utils;
+#if !NET452
+using System.Diagnostics.CodeAnalysis;
+#endif
 
 namespace WpfMath.Atoms
 {
