+++ src/WpfMath.Shared/Data/WpfMathResourceMarker.cs
@@ -1,6 +1,6 @@
 namespace WpfMath.Data;
 
 /// <summary>Marks an assembly containing the WPF-Math resource files (XML and fonts).</summary>
-public static class WpfMathResourceMarker
+internal static class WpfMathResourceMarker
 {
 }
