+++ src/WpfMath.Shared/Rendering/IBrushFactory.cs
@@ -4,5 +4,5 @@ namespace WpfMath.Rendering;
 
 public interface IBrushFactory
 {
-    public IBrush FromColor(RgbaColor color);
+    IBrush FromColor(RgbaColor color);
 }
